% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractivePCA.R
\name{plotInteractivePCA}
\alias{plotInteractivePCA}
\title{Returns an interactive PCA plot.}
\usage{
plotInteractivePCA(
  data,
  meta,
  group,
  model = c("2D", "3D"),
  title = "",
  colors = c(),
  inverse = c(FALSE, FALSE, FALSE)
)
}
\arguments{
\item{data}{numerical matrix or data frame, raw, normalized or transformed. Samples must be in rows and variables in columns}

\item{meta}{data frame containing meta data}

\item{group}{meta data columns to be represented in the plot}

\item{model}{plots a "2D" (default) or "3D" version}

\item{title}{plot's title}

\item{colors}{character vector of custom colors}

\item{inverse}{boolean vector to decide which component plot in reverse, the order is PC1, PC2, PC3}
}
\value{
plotly object containing plot data
}
\description{
Returns an interactive PCA plot.
}
\examples{
Get the result and store it in an object
x <- plotInteractivePCA(combinedData, metaDF, c('Treatment', 'Week'), title="Treatment-week interaction")

direct plot with custom color
plotInteractivePCA(combinedData, metaDF, c('Treatment'), title="Treatment", colors = c("#F8766D", "#00BFC4", "#56B4E9"))

reverse PC1
plotInteractivePCA(combinedData, metaDF, c('Treatment', 'Week'), title="Treatment-week interaction", inverse = c(T,F,F))

3D version
plotInteractivePCA(combinedData, metaDF, c('Treatment', 'Week'), title="Treatment-week interaction", inverse = c(T,F,F), model="3D")
}
